server/requirements.txt
nats.py==2.2.0
python-dotenv==1.0.0